fenrir_vk_2.80
fenrir_vk_2.80

+++ app/src/main/java/dev/ragnarok/fenrir/view/seek/DefaultTimeBar.java
@@ -680,7 +680,7 @@ public class DefaultTimeBar extends View implements TimeBar {
                 if (scrubbing) {
                     if (y < fineScrubYThreshold) {
                         int relativeX = x - lastCoarseScrubXPosition;
-                        positionScrubber(lastCoarseScrubXPosition + relativeX / FINE_SCRUB_RATIO);
+                        positionScrubber(lastCoarseScrubXPosition + (float) relativeX / FINE_SCRUB_RATIO);
                     } else {
                         lastCoarseScrubXPosition = x;
                         positionScrubber(x);
